# Merge performance and serialization changes