--- a/numerichistogram.go
+++ b/numerichistogram.go
@@ -25,21 +25,25 @@ func NewHistogram(n int) *NumericHistogram {
 	}
 }
 
+func (h *NumericHistogram) Reset() {
+	h.total = 0
+	h.bins = h.bins[:1]
+}
+
 func (h *NumericHistogram) Add(n float64) {
 	h.total++
 	for i, bini := range h.bins {
 		if bini.value == n {
 			h.bins[i].count++
-			h.trim()
 			return
 		}
 
 		if bini.value > n {
-			newbin := bin{value: n, count: 1}
+			// Insert a bin at i
 			tail := h.bins[i:]
 			h.bins = h.bins[0 : len(h.bins)+1]
 			copy(h.bins[i+1:], tail)
-			h.bins[i] = newbin
+			h.bins[i] = bin{value: n, count: 1}
 
 			h.trim()
 			return
@@ -144,11 +148,13 @@ func (h *NumericHistogram) trim() {
 				totalCount, // weighted average
 			count: totalCount, // summed heights
 		}
+
+		// Now remove bin minDeltaIndex by replacing minDeltaIndex-1 with mergedbin
 		head := h.bins[0:minDeltaIndex]
 		tail := h.bins[minDeltaIndex+1:]
 		head[minDeltaIndex-1] = mergedbin
 		copy(h.bins[minDeltaIndex:], tail)
-		h.bins = h.bins[:h.maxbins]
+		h.bins = h.bins[:len(h.bins)-1]
 	}
 }