My main contribution is a benchmark to quantify the proposed performance enhancements by @nemothekid

```
Before:
< go test -run none -bench . -benchtime 2s
PASS
BenchmarkNumericHistogram20-8    3000000               891 ns/op
BenchmarkNumericHistogram50-8    2000000              1396 ns/op
BenchmarkNumericHistogram100-8   1000000              2120 ns/op
ok      github.com/signalfx/gohistogram 9.901s

After:
< go test -run none -bench . -benchtime 2s
PASS
BenchmarkNumericHistogram20-8   20000000               205 ns/op
BenchmarkNumericHistogram50-8   10000000               340 ns/op
BenchmarkNumericHistogram100-8   5000000               482 ns/op
ok      github.com/signalfx/gohistogram 10.964s
```