@cep21 2015-12-11
I'm now at:

```
    < go test -run none -bench . -benchtime 2s
    PASS
    BenchmarkNumericHistogram20-8 20000000       137 ns/op
    BenchmarkNumericHistogram50-8 10000000       258 ns/op
    BenchmarkNumericHistogram100-810000000       420 ns/op
```