@Preetam 2017-07-15
@nemosupremo @cep21 Sorry, totally missed this PR. I know it's been a while... any chance you could resolve the conflicts? Otherwise I'll pull your commits into a separate PR and finish things up myself (assuming there are no other problems; I haven't reviewed the code yet).